refactor(Lodeom): rename lodeom.ts as utils.ts
