+++ site/source/pages/simulateurs/lodeom/components/RémunérationInput.tsx
@@ -1,6 +1,6 @@
 import { MontantField } from '@/design-system'
 import { euros } from '@/domaine/Montant'
-import { rémunérationBruteDottedName } from '@/pages/simulateurs/lodeom/lodeom'
+import { rémunérationBruteDottedName } from '@/pages/simulateurs/lodeom/utils'
 
 type Props = {
 	index: number
