+++ site/source/pages/simulateurs/lodeom/utils.ts
